requirements.txt
pymongo==4.4.1
Flask==2.2.2
python-dotenv==1.0.0
PyJWT==2.7.0
cerberus==1.3.5
bcrypt==4.0.0